Revert "FOGL-2409: Always use double data type for trigger_value in OverMaxRule embedded rule"
This reverts commit e2b334a2632fa285a20a9e8d0b8cf6542293ecf1.

+++ C/services/common/overmax_rule.cpp
@@ -350,7 +350,7 @@ bool OverMaxRule::evalAsset(const Value& assetValue,
 			// Check configuration datapoint type
 			if ((*it)->getData().getType() == DatapointValue::T_FLOAT)
 			{
-				assetEval = checkLimit(point, (*it)->getData().toDouble());
+				assetEval = checkLimit(point, (*it)->getData().toInt());
 			}
 			else
 			{
@@ -406,7 +406,7 @@ void OverMaxRule::configure(const ConfigCategory& config)
 
 		if (config.itemExists("trigger_value"))
 		{
-			double maxVal = atof(config.getValue("trigger_value").c_str());
+			double maxVal = atol(config.getValue("trigger_value").c_str());
 			DatapointValue value(maxVal);
 			Datapoint* point = new Datapoint(dataPointName, value);
 			RuleTrigger* pTrigger = new RuleTrigger(dataPointName, point);
